requirements.txt
fastapi[standard]
pydantic-settings
sqlmodel
psycopg2
asyncpg